CPS-544: Create config file while data setup
