+++ gulp-tasks/utils/configs.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const throwError = require('./throw-error.js');
 
 var BACKSTOP_DIR = '.';
 var FILES = {
@@ -11,7 +12,8 @@ var FILES = {
 module.exports = {
   BACKSTOP_DIR: BACKSTOP_DIR,
   FILES: FILES,
-  getSiteConfig: getSiteConfig
+  getSiteConfig: getSiteConfig,
+  touchSiteConfigFile: touchSiteConfigFile
 };
 
 /**
@@ -21,4 +23,21 @@ module.exports = {
  */
 function getSiteConfig () {
   return JSON.parse(fs.readFileSync(FILES.siteConfig));
+}
+
+/**
+ * Creates the site config file is in the backstopjs folder, if it doesn't exists yet
+ */
+function touchSiteConfigFile () {
+  try {
+    fs.readFileSync(FILES.siteConfig);
+  } catch (err) {
+    fs.copyFileSync(FILES.siteConfigSample, FILES.siteConfig);
+
+    throwError(
+      'No site-config.json file detected!\n' +
+      '\tOne has been created for you \n' +
+      '\tPlease insert the real value for each placeholder and try again'
+    );
+  }
 }
