+++ package.json
@@ -19,6 +19,14 @@
     "moment": "^2.24.0",
     "path": "^0.12.7",
     "plugin-error": "^1.0.1",
-    "yargs": "^14.2.0"
+    "yargs": "^14.2.0",
+    "eslint": "^6.5.1",
+    "eslint-config-semistandard": "^15.0.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsdoc": "^15.11.1",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1"
   }
 }
